refactor: move game over listener logic to its own class
