+++ src/main/kotlin/com/github/frozensync/games/shiritori/DisqualificationTimer.kt
@@ -3,20 +3,19 @@ package com.github.frozensync.games.shiritori
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
-import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 
 internal class DisqualificationTimer(
     private val time: Long, // ms
     private val shiritoriRepository: ShiritoriRepository,
-    private val notificationChannel: Channel<Shiritori>,
+    private val gameOverListener: GameOverListener,
 ) {
     private val action: suspend (Long) -> Unit = { id ->
         shiritoriRepository.findById(id)
             ?.removePlayer()
             ?.let {
                 shiritoriRepository.save(it)
-                notificationChannel.send(it)
+                gameOverListener.sendGameOver(it)
             }
     }
 
@@ -29,4 +28,7 @@ internal class DisqualificationTimer(
             action.invoke(id)
         }
     }
+
+    override fun toString(): String =
+        "DisqualificationTimer{time=$time}"
 }
